type NodeCache struct
